requirements.txt
Flask==3.0.3
gunicorn==23.0.0
mysql-connector-python==9.0.0
numpy==2.1.1
